+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Core: enter key on grid did not open edit dialog
+- Tasks: Sorting by 'responsible' and 'categories' works
 
 15-03-2022 6.6.54
 - Core: when using SMIME signing in the config.php it failed on a certificate failure